+++ src/lib.rs
@@ -3,3 +3,9 @@ pub mod parsers;
 
 /// search contains functions for searching files and tags.
 pub mod search;
+
+/// interactive inspection ui
+pub mod inspect;
+
+/// system commands on files
+pub mod commands;
