+++ src/main.rs
@@ -1,5 +1,4 @@
 use std::io::{stdout, BufRead, IsTerminal};
-use std::{path::Path, process::Command};
 
 use colored::Colorize;
 use crossterm::terminal::{
@@ -8,6 +7,8 @@ use crossterm::terminal::{
 use crossterm::ExecutableCommand;
 use pest::Parser;
 
+use tag::commands::{execute_command_on_file, execute_filter_command_on_file};
+use tag::inspect;
 use tag::search::TaggedFile;
 use tag::{
     parsers::searchquery::{construct_query_ast, evaluate_ast, QueryParser, Rule},
@@ -61,60 +62,6 @@ mod cli {
     }
 }
 
-fn execute_command_on_file(path: &Path, command: &str) -> String {
-    let command = command.replace("#FILE#", path.to_str().unwrap());
-
-    let output = if cfg!(target_os = "windows") {
-        Command::new("cmd").arg("/C").arg(command.clone()).output()
-    } else {
-        Command::new("bash").arg("-c").arg(command.clone()).output()
-    };
-
-    if let Err(e) = &output {
-        eprintln!(
-            "{} Wasn't able to execute command {}: {}",
-            "[ERROR]".red().bold(),
-            command.blue().underline(),
-            e.to_string().red()
-        );
-    }
-
-    let output = output.unwrap();
-    let output_string = std::str::from_utf8(output.stdout.as_slice());
-
-    if let Err(e) = &output_string {
-        eprintln!(
-            "{} Failed to get output from command {}: {}",
-            "[ERROR]".red().bold(),
-            command.blue().underline(),
-            e.to_string().red()
-        );
-    }
-
-    output_string.unwrap().to_string()
-}
-
-fn execute_filter_command_on_file(path: &Path, command: &str) -> bool {
-    let command = command.replace("#FILE#", path.to_str().unwrap());
-
-    let output = if cfg!(target_os = "windows") {
-        Command::new("cmd").arg("/C").arg(command.clone()).output()
-    } else {
-        Command::new("bash").arg("-c").arg(command.clone()).output()
-    };
-
-    if let Err(e) = &output {
-        eprintln!(
-            "{} Wasn't able to execute command {}: {}",
-            "[ERROR]".red().bold(),
-            command.blue().underline(),
-            e.to_string().red()
-        );
-    }
-
-    output.unwrap().status.success()
-}
-
 fn non_interactive_output(file: &TaggedFile, command_output: &str) {
     println!("\t{}", format!("tags: {:?}", file.tags).blue());
 
@@ -123,289 +70,6 @@ fn non_interactive_output(file: &TaggedFile, command_output: &str) {
             "\tOutput of command:\n{}",
             textwrap::indent(command_output, "\t\t")
         );
-    }
-}
-
-mod interactive_output {
-    use crossterm::event;
-    use crossterm::event::{Event, KeyCode};
-    use itertools::Itertools;
-    use ratatui::backend::CrosstermBackend;
-    use ratatui::layout::{Alignment, Constraint, Direction, Layout, Rect};
-    use ratatui::style::{Style, Stylize};
-    use ratatui::text::{Line, Span};
-    use ratatui::widgets::{Block, Borders, Clear, Paragraph, Tabs, Wrap};
-    use ratatui::{symbols, Frame, Terminal};
-    use std::io::{self, stdout};
-    use std::rc::Rc;
-    use tui_textarea::{Input, Key, TextArea};
-
-    use tag::search::TaggedFile;
-
-    use crate::execute_command_on_file;
-
-    /// `InteractiveInputs` contains possible inputs for interactive mode.
-    #[derive(Default)]
-    struct InteractiveInputs {
-        pub tab_index: usize,
-        pub file_index: usize,
-        pub scroll_index: u16,
-        pub command_mode: bool,
-        pub quit: bool,
-    }
-
-    pub fn interactive_output(files: &[TaggedFile], command_outputs: &[String]) -> io::Result<()> {
-        let mut terminal = Terminal::new(CrosstermBackend::new(stdout()))?;
-
-        // the command textarea
-        let mut textarea = TextArea::default();
-        textarea.set_cursor_line_style(Style::default());
-        textarea.set_placeholder_text("Enter a command");
-        textarea.set_block(
-            Block::new()
-                .title("command")
-                .borders(Borders::all())
-                .border_style(Style::default().red().on_black())
-                .style(Style::default().black().on_white()),
-        );
-
-        let mut interactive_inputs = InteractiveInputs::default();
-        while !interactive_inputs.quit {
-            let file = &files[interactive_inputs.file_index];
-            let command_output = command_outputs[interactive_inputs.file_index].clone();
-
-            terminal.draw(|frame| {
-                interactive_output_ui(
-                    file,
-                    command_output.as_str(),
-                    &mut interactive_inputs,
-                    &mut textarea,
-                    frame,
-                );
-            })?;
-            interactive_inputs = handle_events(&interactive_inputs)?;
-
-            // prevent an overflow of the index
-            // and also handle wrapping
-            interactive_inputs.tab_index %= 3;
-            interactive_inputs.file_index %= files.len();
-        }
-
-        Ok(())
-    }
-
-    fn interactive_output_ui(
-        file: &TaggedFile,
-        command_output: &str,
-        interactive_inputs: &mut InteractiveInputs,
-        text_area: &mut TextArea,
-        frame: &mut Frame,
-    ) {
-        let area = layout(frame.size(), Direction::Vertical, &[1, 0, 1]);
-
-        render_tabs(area[0], frame, interactive_inputs);
-
-        render_tab_content(
-            file,
-            command_output,
-            interactive_inputs.tab_index,
-            interactive_inputs.scroll_index,
-            area[1],
-            frame,
-        );
-
-        render_help_menu(area[2], frame);
-
-        if interactive_inputs.command_mode {
-            interactive_inputs.command_mode = command_mode_input(file, text_area).unwrap();
-            command_mode_render(text_area, frame);
-        }
-    }
-
-    fn render_tabs(area: Rect, frame: &mut Frame, interactive_inputs: &InteractiveInputs) {
-        let tabs = Tabs::new(vec!["File Content", "Command Output", "Tags"])
-            .style(Style::default().white())
-            .highlight_style(Style::default().blue())
-            .select(interactive_inputs.tab_index)
-            .divider(symbols::DOT);
-
-        frame.render_widget(tabs, area);
-    }
-
-    fn command_mode_render(text_area: &mut TextArea, frame: &mut Frame) {
-        let layout =
-            Layout::default().constraints([Constraint::Length(3), Constraint::Min(1)].as_slice());
-
-        let area = Rect::new(0, frame.size().height / 2, frame.size().width, 10);
-
-        frame.render_widget(Clear, layout.split(area)[0]);
-        frame.render_widget(text_area.widget(), layout.split(area)[0]);
-    }
-
-    fn command_mode_input(
-        file: &TaggedFile,
-        text_area: &mut TextArea,
-    ) -> Result<bool, std::io::Error> {
-        match crossterm::event::read()?.into() {
-            Input { key: Key::Esc, .. } => {
-                return Ok(false);
-            }
-            Input {
-                key: Key::Enter, ..
-            } => {
-                execute_command_on_file(&file.path, &text_area.lines()[0]);
-                return Ok(false);
-            }
-            Input {
-                key: Key::Char('m'),
-                ctrl: true,
-                ..
-            } => {}
-            input => {
-                text_area.input(input);
-            }
-        }
-
-        Ok(true)
-    }
-
-    fn render_tab_content(
-        file: &TaggedFile,
-        command_output: &str,
-        tab_index: usize,
-        scroll_index: u16,
-        area: Rect,
-        frame: &mut Frame,
-    ) {
-        let content = match tab_index {
-            0 => std::fs::read_to_string(&file.path).unwrap(),
-            1 => command_output.to_string(),
-            2 => {
-                let mut out_string = String::new();
-                for tag in &file.tags {
-                    out_string += tag.as_str();
-                    out_string.push('\n');
-                }
-                out_string
-            }
-            _ => unreachable!(), // tabs are constrained to be between 0 and 2
-        };
-
-        #[allow(clippy::cast_possible_truncation)]
-        let scroll_index = scroll_index % content.lines().collect_vec().len() as u16;
-
-        let paragraph = Paragraph::new(content)
-            .block(
-                Block::new()
-                    .title(file.path.to_str().unwrap())
-                    .borders(Borders::all()),
-            )
-            .style(Style::new())
-            .alignment(Alignment::Left)
-            .wrap(Wrap { trim: false })
-            .scroll((scroll_index, 0));
-
-        frame.render_widget(paragraph, area);
-    }
-
-    fn render_help_menu(area: Rect, frame: &mut Frame) {
-        let keys = [
-            ("q", "Quit"),
-            ("Up-Arrow/k", "Scroll Up"),
-            ("Down-Arrow/j", "Scroll Down"),
-            ("n", "Next File"),
-            ("p", "Previous File"),
-            ("Tab/Right-Arrow/l", "Next Tab"),
-            ("Shift+Tab/Left-Arrow/h", "Previous Tab"),
-            ("c", "Execute a command"),
-        ];
-
-        let spans = keys
-            .iter()
-            .flat_map(|(key, desc)| {
-                let key = Span::styled(format!("| {key} "), Style::new().green());
-                let desc = Span::styled(format!(" {desc} | "), Style::new().green());
-                [key, desc]
-            })
-            .collect_vec();
-
-        let paragraph = Paragraph::new(Line::from(spans)).alignment(Alignment::Center);
-
-        frame.render_widget(paragraph, area);
-    }
-
-    /// simple helper method to split an area into multiple sub-areas
-    /// copied and slightly modified from
-    /// [here](https://docs.rs/ratatui/latest/src/demo2/root.rs.html#34)
-    pub fn layout(area: Rect, direction: Direction, heights: &[u16]) -> Rc<[Rect]> {
-        let constraints = heights
-            .iter()
-            .map(|&h| {
-                if h > 0 {
-                    Constraint::Length(h)
-                } else {
-                    Constraint::Min(0)
-                }
-            })
-            .collect_vec();
-        Layout::default()
-            .direction(direction)
-            .constraints(constraints)
-            .split(area)
-    }
-
-    fn handle_events(previous_inputs: &InteractiveInputs) -> io::Result<InteractiveInputs> {
-        let mut interactive_inputs = InteractiveInputs {
-            tab_index: previous_inputs.tab_index,
-            file_index: previous_inputs.file_index,
-            scroll_index: previous_inputs.scroll_index,
-            command_mode: previous_inputs.command_mode,
-            ..Default::default()
-        };
-
-        if event::poll(std::time::Duration::from_millis(50))? {
-            if let Event::Key(key) = event::read()? {
-                if key.kind != event::KeyEventKind::Press {
-                    return Ok(interactive_inputs);
-                }
-
-                match key.code {
-                    KeyCode::Char('n') => interactive_inputs.file_index += 1,
-                    KeyCode::Char('p') => {
-                        if interactive_inputs.file_index == 0 {
-                            interactive_inputs.file_index = usize::MAX;
-                        } else {
-                            interactive_inputs.file_index -= 1;
-                        }
-                    }
-                    KeyCode::Char('l') | KeyCode::Right | KeyCode::Tab => {
-                        interactive_inputs.tab_index += 1;
-                    }
-                    KeyCode::Char('h') | KeyCode::Left | KeyCode::BackTab => {
-                        interactive_inputs.tab_index -= 1;
-                    }
-                    KeyCode::Char('k') | KeyCode::Up => {
-                        if interactive_inputs.scroll_index == 0 {
-                            interactive_inputs.scroll_index = u16::MAX;
-                        } else {
-                            interactive_inputs.scroll_index -= 1;
-                        }
-                    }
-                    KeyCode::Char('j') | KeyCode::Down => {
-                        if interactive_inputs.scroll_index == u16::MAX {
-                            interactive_inputs.scroll_index = u16::MIN;
-                        } else {
-                            interactive_inputs.scroll_index += 1;
-                        }
-                    }
-                    KeyCode::Char('c') => interactive_inputs.command_mode = true,
-                    KeyCode::Char('q') => interactive_inputs.quit = true,
-                    _ => return Ok(interactive_inputs),
-                }
-            }
-        }
-
-        Ok(interactive_inputs)
     }
 }
 
@@ -505,7 +169,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     }
 
     if args.inspect {
-        interactive_output::interactive_output(&file_matched_index, &command_outputs)?;
+        inspect::interactive_output(&file_matched_index, &command_outputs)?;
 
         disable_raw_mode()?;
         stdout().execute(LeaveAlternateScreen)?;
